examples/quick_start.py
"""
Quick Start Example for GIIM

This example demonstrates the intended API usage for GIIM model training and evaluation.
It shows how to initialize components, load data, train the model, and evaluate performance.

⚠️ IMPLEMENTATION PENDING - This is an API demonstration for documentation purposes.
Full implementation will be released following institutional review.

Usage:
    python examples/quick_start.py
"""

import sys
from pathlib import Path
import yaml

# Add parent directory to path
sys.path.insert(0, str(Path(__file__).parent.parent))


def main():
    """
    Quick start example demonstrating the GIIM API.
    
    This example shows the intended workflow for:
    1. Loading configuration
    2. Initializing all components
    3. Loading data
    4. Creating and training the model
    5. Evaluating performance
    
    ⚠️ This is a demonstration of the planned API structure.
    Actual implementation is pending institutional review.
    """
    
    print("=" * 80)
    print("GIIM Quick Start Example")
    print("=" * 80)
    print(f"\n⚠️  IMPLEMENTATION PENDING")
    print(f"This example demonstrates the intended API usage.")
    print(f"Full implementation will be released following institutional review.\n")
    print("=" * 80)
    
    # ====================
    # 1. Load Configuration
    # ====================
    print("\n📋 Step 1: Load Configuration")
    print("-" * 80)
    
    config_path = Path(__file__).parent.parent / "configs" / "default.yaml"
    print(f"Loading config from: {config_path}")
    
    if not config_path.exists():
        print(f"❌ Config file not found: {config_path}")
        return False
    
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    
    print(f"\n✅ Configuration loaded:")
    print(f"  Dataset: {config['dataset']['name']}")
    print(f"  Views: {config['dataset']['views'][config['dataset']['name']]}")
    print(f"  Model architecture:")
    print(f"    - Input dim: {config['model']['in_dim']}")
    print(f"    - Hidden dims: {config['model']['hidden_dims']}")
    print(f"    - Output classes: {config['model']['num_classes']}")
    print(f"  Training:")
    print(f"    - Learning rate: {config['training']['learning_rate']}")
    print(f"    - Batch size: {config['training']['batch_size']}")
    print(f"    - Epochs: {config['training']['epochs']}")
    
    # ====================
    # 2. Initialize Components
    # ====================
    print("\n🔧 Step 2: Initialize Components")
    print("-" * 80)
    
    print("\nIntended initialization code:")
    print("""
from giim import (
    DatasetLoader,
    FeatureExtractor,
    GraphBuilder,
    GIIMModel,
    Utils,
    Trainer,
    Evaluation
)

# Initialize data loader
dataset_loader = DatasetLoader(config)

# Initialize feature extractor (ConvNeXt-Tiny)
feature_extractor = FeatureExtractor(config)

# Initialize graph builder
graph_builder = GraphBuilder(config['dataset']['name'])

# Initialize utility functions for imputation
utils = Utils(config, dataset_loader)
    """)
    
    print("⚠️  Component implementations are pending institutional review.")
    
    # ====================
    # 3. Create Model
    # ====================
    print("\n🧠 Step 3: Create GIIM Model")
    print("-" * 80)
    
    print("\nIntended model creation code:")
    print("""
model = GIIMModel(
    in_dim=config['model']['in_dim'],           # 769 (768 + 1 bias)
    hidden_dims=config['model']['hidden_dims'], # [512, 256, 128, 64]
    num_classes=config['model']['num_classes'], # 4 for liver CT
    num_views=config['model']['num_views']      # 3 for liver CT
)

print(f"Model created with {sum(p.numel() for p in model.parameters())} parameters")
    """)
    
    print("⚠️  Model implementation is pending institutional review.")
    
    # ====================
    # 4. Load Data
    # ====================
    print("\n📊 Step 4: Load Data")
    print("-" * 80)
    
    print("\nIntended data loading code:")
    print("""
# Load train/val/test splits
train_data = dataset_loader.load_data(dataset_name, "train")
val_data = dataset_loader.load_data(dataset_name, "val")
test_data = dataset_loader.load_data(dataset_name, "test")

print(f"Training samples: {len(train_data)}")
print(f"Validation samples: {len(val_data)}")
print(f"Test samples: {len(test_data)}")

# Example patient data structure:
# PatientData(
#     patient_id='patient_001',
#     labels={'class': 2},  # Malignant
#     views={
#         'arterial': [tensor(...), tensor(...)],  # List of ROI tensors
#         'venous': [tensor(...), tensor(...)],
#         'delay': [tensor(...), tensor(...)]
#     },
#     missing_views=set()  # No missing views
# )
    """)
    
    print("⚠️  Data loading implementation is pending institutional review.")
    print("\nExpected data structure:")
    print("  data/")
    print("    liver_ct/")
    print("      ├── train.csv")
    print("      ├── val.csv")
    print("      ├── test.csv")
    print("      └── images/")
    print("          ├── patient001_lesion01_arterial.png")
    print("          └── ...")
    print("\nSee docs/datasets.md for data preparation guidelines.")
    
    # ====================
    # 5. Train Model
    # ====================
    print("\n🎓 Step 5: Train Model")
    print("-" * 80)
    
    print("\nIntended training code:")
    print("""
# Initialize trainer
trainer = Trainer(
    model=model,
    train_loader=train_data,
    val_loader=val_data,
    config=config,
    feature_extractor=feature_extractor,
    graph_builder=graph_builder,
    utils=utils
)

# Train with early stopping
history = trainer.train()

print(f"Training completed!")
print(f"Best validation loss: {min(history['val_losses']):.4f}")
print(f"Best model saved at: {history['best_model_path']}")
    """)
    
    print("⚠️  Training implementation is pending institutional review.")
    
    # ====================
    # 6. Evaluate Model
    # ====================
    print("\n📈 Step 6: Evaluate Model")
    print("-" * 80)
    
    print("\nIntended evaluation code:")
    print("""
# Initialize evaluator
evaluation = Evaluation(
    model=model,
    test_loader=test_data,
    config=config,
    feature_extractor=feature_extractor,
    graph_builder=graph_builder,
    utils=utils
)

# Run comprehensive evaluation
results = evaluation.evaluate()

# Print results table (reproduces paper Table 3)
evaluation.print_summary(results)

# Example output:
# ┌────────────┬──────────────┬──────────┬─────────┬─────────┬──────────┐
# │ Test Set   │ Missing Rate │ Method   │ Accuracy│ AUC     │ F1-Score │
# ├────────────┼──────────────┼──────────┼─────────┼─────────┼──────────┤
# │ Full-view  │ 0.0          │ RAG      │ 0.852   │ 0.912   │ 0.847    │
# │ Full-view  │ 0.5          │ RAG      │ 0.785   │ 0.863   │ 0.772    │
# │ Miss-view  │ 0.0          │ RAG      │ 0.789   │ 0.854   │ 0.776    │
# └────────────┴──────────────┴──────────┴─────────┴─────────┴──────────┘
    """)
    
    print("⚠️  Evaluation implementation is pending institutional review.")
    
    # ====================
    # Summary
    # ====================
    print("\n" + "=" * 80)
    print("📝 API Summary")
    print("=" * 80)
    print("""
Complete Workflow:
1. Load configuration from YAML file
2. Initialize components:
   - DatasetLoader: Load and preprocess medical images
   - FeatureExtractor: ConvNeXt-based feature extraction
   - GraphBuilder: Heterogeneous graph construction
   - Utils: Missing view imputation methods
3. Create GIIM model with specified architecture
4. Load train/val/test data
5. Train model with Trainer class
6. Evaluate with Evaluation class across all scenarios

For More Information:
- Architecture details: docs/architecture.md
- Dataset preparation: docs/datasets.md
- Configuration options: configs/default.yaml
    """)
    
    print("=" * 80)
    print("⚠️  Implementation pending institutional review")
    print("See README.md for project status and expected release timeline")
    print("=" * 80)
    
    return True


if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
