requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.0.0

# Feature extraction
timm>=0.6.0

# Data processing
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
pillow>=9.0.0

# Configuration
pyyaml>=6.0

# Visualization and logging
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.10.0
tqdm>=4.64.0

# Utilities
scipy>=1.7.0

